03-readme-generator/examples/sample_project/requirements.txt
click>=8.0
pyyaml>=6.0
rich>=13.0
